moved test files to the lib folder
