+++ lib/atat-net-infra-pipeline-stage.ts
@@ -35,7 +35,10 @@ export class NetInfraPipelineStage extends cdk.Stage {
     if (props.environmentName === 'Dev') {
       const atatFirewallLoadBalancer = new AlbStack(this, 'AtatALB', {
         atatfirewallVpc: atatFirewallVpc,
-        apiDomain: props.apiDomain
+        apiDomain: props.apiDomain,
+        env: {
+          region: this.region,
+        },
       })
     }
